stop running test cases when related to known bugs
Stops running test cases that fail or error due to known limitations (while still allowing to easily re-include them, see test/helper.rb). This makes it way easier to refactor or fix things as the test suite can be mostly green and it's more easy to see which errors might be new.